definition and declaration must match
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@17155 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/other/ftgl/src/FTVectoriser.cpp
@@ -117,7 +117,7 @@ void FTMesh::End()
 }
 
 
-const FTTesselation* const FTMesh::Tesselation(unsigned int index) const
+const FTTesselation* const FTMesh::Tesselation(size_t index) const
 {
     return (index < tesselationList.size()) ? tesselationList[index] : NULL;
 }
